requirements.txt
psycopg2-binary
python-dotenv
flask
flask-cors
pytest
